+++ cm/hal_ioreg.c
@@ -1,10 +1,12 @@
 
 #include "hal_ioreg.h"
 
-u32 hal_IOReg_Read(u32 port) {
+
+
+u8 hal_IOReg_Read(u8 port) {
     return bsl_IOReg_Read(port);
 }
 
-void hal_IOReg_Write(u32 port, u32 value) {
+void hal_IOReg_Write(u8 port, u8 value) {
     bsl_IOReg_Write(port, value);
 }
